refactor: enforce props types
