+++ packages/shared/src/modals/index.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'react'
 import { SingletonModal } from '@masknet/shared-base'
 import { SelectProviderModal, type SelectProviderDialogOpenProps } from './SelectProviderDialog/index.js'
-import { WalletConnectQRCode, type WalletConnectQRCodeOpenProps } from './WalletConnectQRCodeDialog/index.js'
+import { WalletConnectQRCodeModal, type WalletConnectQRCodeOpenProps } from './WalletConnectQRCodeDialog/index.js'
 import { WalletRiskWarningModal, type WalletRiskWarningModalOpenProps } from './WalletRiskWarningDialog/index.js'
 import { WalletStatusModal, type WalletStatusModalOpenProps } from './WalletStatusDialog/index.js'
 import { LeavePageConfirmModal, type LeavePageConfirmDialogOpenProps } from './LeavePageConfirmDialog/index.js'
@@ -15,7 +15,7 @@ export const LeavePageConfirmDialog = new SingletonModal<LeavePageConfirmDialogO
 export const Modals = memo(function Modals() {
     return (
         <>
-            <WalletConnectQRCode ref={WalletConnectQRCodeDialog.register} />
+            <WalletConnectQRCodeModal ref={WalletConnectQRCodeDialog.register} />
             <SelectProviderModal ref={SelectProviderDialog.register} />
             <WalletStatusModal ref={WalletStatusDialog.register} />
             <WalletRiskWarningModal ref={WalletRiskWarningDialog.register} />
